python/requirements.txt
jax[cpu]>=0.6.0
jaxlib>=0.6.0
jaxtyping>=0.3.0
flax>=0.10.7
numpy>=1.20.0
attrs>=25.0.0
chex>=0.1.89
yapf>=0.43.0
absl-py>=2.2.0
flatbuffers>=25.0.0
packaging>=25.0.0